chore: rename xsum files
